news_aggregator/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
transformers==4.35.0
torch==2.1.0
newspaper3k==0.2.8
feedparser==6.0.10
textblob==0.17.1
sumy==0.11.0
nltk==3.8.1
lxml==4.9.3
gunicorn==21.2.0
